refactor(renderer): extract the cascaded shadow pass into shadow_pass.rs
Render-graph migration cluster 4 prep: cascade fitting, the
ticket-004 cache-hit skip, and the per-cascade depth renders become
record_shadow_pass (199 lines out of end_frame_with_scene). Goldens —
including the shadowed lit_primitives_3d scene — pixel-identical.
mod.rs ratchets 13108 -> 12914.
